@@ -1,16 +1,16 @@
-import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
+import {NgModule} from "@angular/core";
+import {Routes, RouterModule} from "@angular/router";
 
-import { HomeComponent } from './home/home.component';
-import { NotesComponent } from './notes/notes.component';
+import {HomeComponent} from "./home/home.component";
+import {NotesComponent} from "./notes/notes.component";
 
 const routes: Routes = [
-  { path: '', component: HomeComponent },
-  { path: 'notes', component: NotesComponent },
+  {path: "", component: HomeComponent},
+  {path: "notes", component: NotesComponent}
 ];
 
 @NgModule({
   imports: [RouterModule.forRoot(routes)],
   exports: [RouterModule]
 })
-export class AppRoutingModule { }
+export class AppRoutingModule {}
