@@ -1,15 +1,15 @@
-import { INote } from '../shared/interfaces';
+import {INote} from "../shared/interfaces";
 
 export class Note implements INote {
-    id: number;
-    Title: string;
-    content: string;
-    favorite?: boolean;
+  id: number;
+  Title: string;
+  content: string;
+  favorite?: boolean;
 
-    constructor(id: number, title: string, content: string) {
-        this.id = id;
-        this.Title = title;
-        this.content = content;
-        this.favorite = false;
-    }
+  constructor(id: number, title: string, content: string) {
+    this.id = id;
+    this.Title = title;
+    this.content = content;
+    this.favorite = false;
+  }
 }
