@@ -1,16 +1,15 @@
-import { Component, OnInit, Inject } from '@angular/core';
+import {Component, OnInit, Inject} from "@angular/core";
 import {INote} from "../shared/interfaces";
 import {DataService} from "../core/data.service";
 import {MatDialog, MatDialogRef, MAT_DIALOG_DATA} from "@angular/material";
-import {NoteAddComponent} from '../notes/notes.component'
+import {NoteAddComponent} from "../notes/notes.component";
 
 @Component({
-  selector: 'app-home',
-  templateUrl: './home.component.html',
-  styleUrls: ['./home.component.scss']
+  selector: "app-home",
+  templateUrl: "./home.component.html",
+  styleUrls: ["./home.component.scss"]
 })
 export class HomeComponent implements OnInit {
-
   notes: INote[] = [];
 
   noteData: any = {
@@ -19,7 +18,7 @@ export class HomeComponent implements OnInit {
     index: 0
   };
 
-  constructor(private dataService: DataService, public dialog: MatDialog) { }
+  constructor(private dataService: DataService, public dialog: MatDialog) {}
 
   ngOnInit() {
     this.notes = this.dataService.getNotes();
@@ -27,7 +26,10 @@ export class HomeComponent implements OnInit {
 
   private createNote(): void {
     if (this.noteData.newtitle !== "" && this.noteData.newcontent !== "") {
-      this.dataService.addNote(this.noteData.newtitle, this.noteData.newcontent);
+      this.dataService.addNote(
+        this.noteData.newtitle,
+        this.noteData.newcontent
+      );
       this.update();
     }
     // set the model values to '' again
@@ -42,8 +44,7 @@ export class HomeComponent implements OnInit {
   openAddDialog(): void {
     const dialogRef = this.dialog.open(NoteAddComponent, {
       width: "250px",
-      data: {ndata: this.noteData }
-     
+      data: {ndata: this.noteData}
     });
 
     dialogRef.afterClosed().subscribe(result => {
@@ -56,5 +57,3 @@ export class HomeComponent implements OnInit {
     this.openAddDialog();
   }
 }
-
-
